put everything into a weather engine, now can have multiple weather engine meaning multiple cards that update based on different searches
