+++ src/App.css
@@ -1,6 +1,6 @@
 @import url("https://fonts.googleapis.com/css?family=Fira+Sans|Merriweather&display=swap");
 .App {
-  margin: 0 auto;
+  margin: 20 auto;
   color: white;
   display: flex;
   flex-direction: row;
